LayoutTests/fast/dom/title-setter-new-text-node-svg-expected.txt

Tests that setting document.title does not reuse title's text node child

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS svgDocument.title is "aaa"
PASS oldTextNode.textContent is "aaa"
svgDocument.title = 'bbb'
PASS oldTextNode.textContent is "aaa"
PASS oldTextNode is not newTextNode
PASS svgDocument.title is "bbb"
PASS newTextNode.textContent is "bbb"
PASS successfullyParsed is true

TEST COMPLETE
